chore: Remove deprecated HDXLineChart and HDXMultiSeriesTimeChart
