+++ packages/app/src/HDXMultiSeriesTimeChart.tsx
@@ -2,6 +2,7 @@ import { memo, useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import Link from 'next/link';
 import cx from 'classnames';
 import { add, format } from 'date-fns';
+import { withErrorBoundary } from 'react-error-boundary';
 import { toast } from 'react-toastify';
 import {
   Bar,
@@ -26,7 +27,6 @@ import {
   seriesColumns,
   seriesToUrlSearchQueryParam,
 } from './ChartUtils';
-import { HDXLineChartTooltip } from './HDXLineChart';
 import type { ChartSeries, NumberFormat } from './types';
 import useUserPreferences, { TimeFormat } from './useUserPreferences';
 import { formatNumber } from './utils';
@@ -35,6 +35,42 @@ import { semanticKeyedColor, TIME_TOKENS, truncateMiddle } from './utils';
 import styles from '../styles/HDXLineChart.module.scss';
 
 const MAX_LEGEND_ITEMS = 4;
+
+const HDXLineChartTooltip = withErrorBoundary(
+  memo((props: any) => {
+    const timeFormat: TimeFormat = useUserPreferences().timeFormat;
+    const tsFormat = TIME_TOKENS[timeFormat];
+    const { active, payload, label, numberFormat } = props;
+    if (active && payload && payload.length) {
+      return (
+        <div className={styles.chartTooltip}>
+          <div className={styles.chartTooltipHeader}>
+            {format(new Date(label * 1000), tsFormat)}
+          </div>
+          <div className={styles.chartTooltipContent}>
+            {payload
+              .sort((a: any, b: any) => b.value - a.value)
+              .map((p: any) => (
+                <div key={p.dataKey} style={{ color: p.color }}>
+                  {truncateMiddle(p.name ?? p.dataKey, 70)}:{' '}
+                  {numberFormat ? formatNumber(p.value, numberFormat) : p.value}
+                </div>
+              ))}
+          </div>
+        </div>
+      );
+    }
+    return null;
+  }),
+  {
+    onError: console.error,
+    fallback: (
+      <div className="text-danger px-2 py-1 m-2 fs-8 font-monospace bg-danger-transparent">
+        An error occurred while rendering the tooltip.
+      </div>
+    ),
+  },
+);
 
 function CopyableLegendItem({ entry }: any) {
   return (
